新文件：   .idea/dbnavigator.xml 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/__pycache__/ctr_vector_db.cpython-310.pyc" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/__pycache__/model.cpython-310.pyc" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/__pycache__/pdf_split.cpython-310.pyc" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/chat_with_rag.py" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/ctr_vector_db.py" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/model.py" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/routers/__pycache__/chat_retrieval_module.cpython-310.pyc" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/routers/chat_retrieval_module.py" 	修改：     "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/routers/memory_retrieval.py" 	新文件：   "4.\345\274\200\345\217\221/front/demo/demo/1.user/rag/routers/serve_module.py"
